kernel - Fix bug in physio()
* Fix an integer overflow in physio() when attempting to dd a block size
  >= 2G.  An int was being used instead of size_t.

Reported-by: tpchin
Bug: 2981

+++ sys/kern/kern_physio.c
@@ -35,8 +35,8 @@ physio(cdev_t dev, struct uio *uio, int ioflag)
 	int i;
 	int error;
 	int saflags;
-	int iolen;
-	int bcount;
+	size_t iolen;
+	size_t bcount;
 	caddr_t ubase;
 	struct buf *bp;
